requirements.txt
pycparser==2.21
PyYAML==6.0
